connector/code-snippet_2_0_2/requirements.txt
PyYAML
jsonpickle
